Reformat Code

+++ src/BmiService.java
@@ -1,9 +1,9 @@
 public class BmiService {
 
-    public double calculate (int bodyMass, double growth) {
+    public double calculate(int bodyMass, double growth) {
         double degree = growth * growth;
         double bmi = bodyMass / degree;
 
-        return (double) bmi;
+        return bmi;
     }
 }
